clean up fixtures
