+++ tests/conftest.py
@@ -41,7 +41,7 @@ def pytest_sessionfinish(session, exitstatus):
     env_vars.stop()
 
 
-@pytest.fixture(autouse=True)
+@pytest.fixture()
 def runner():
     print("Hiiiiiiiasdufhas8dfoasdijfak")
     # reset CLI config file
@@ -65,3 +65,46 @@ def runner():
     # stop servers
     stop_result = cli_runner.invoke(main, ["server", "stop", "--all"])
     assert stop_result.exit_code == 0, stop_result.output
+
+
+@pytest.fixture(autouse=True)
+def create_bridge(runner):
+    # create bridge
+    create_result = runner.invoke(
+        main,
+        [
+            "bridge",
+            "create",
+            "--name=test_bridge",
+            "--chains",
+            "locking_chain",
+            "issuing_chain",
+            "--witness",
+            "witness0",
+            "--witness",
+            "witness1",
+            "--witness",
+            "witness2",
+            "--witness",
+            "witness3",
+            "--witness",
+            "witness4",
+            "--verbose",
+        ],
+    )
+    assert create_result.exit_code == 0, create_result.output
+
+    config_dir = os.path.abspath(os.getenv("XCHAIN_CONFIG_DIR"))
+    with open(os.path.join(config_dir, "bridge_bootstrap.json")) as f:
+        bootstrap = json.load(f)
+
+    locking_door = bootstrap["locking_chain_door"]["id"]
+    fund_result = runner.invoke(
+        main, ["fund", f"--account={locking_door}", "--chain=locking_chain"]
+    )
+    assert fund_result.exit_code == 0, fund_result.output
+
+    build_result = runner.invoke(
+        main, ["bridge", "build", "--bridge=test_bridge", "--verbose"]
+    )
+    assert build_result.exit_code == 0, build_result.output
